+++ gradle/libs.versions.toml
@@ -13,6 +13,7 @@ composeBom = "2024.06.00"
 composeMaterial = "1.3.1"
 composeFoundation = "1.3.1"
 coreSplashscreen = "1.0.1"
+wearToolingPreview = "1.0.0"
 
 [libraries]
 androidGradlePlugin = { group = "com.android.tools.build", name = "gradle", version.ref = "agp" }
@@ -37,6 +38,7 @@ androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit
 androidx-compose-material = { group = "androidx.wear.compose", name = "compose-material", version.ref = "composeMaterial" }
 androidx-compose-foundation = { group = "androidx.wear.compose", name = "compose-foundation", version.ref = "composeFoundation" }
 androidx-core-splashscreen = { group = "androidx.core", name = "core-splashscreen", version.ref = "coreSplashscreen" }
+androidx-wear-tooling-preview = { group = "androidx.wear", name = "wear-tooling-preview", version.ref = "wearToolingPreview" }
 
 [plugins]
 androidGradlePlugin = { id = "com.android.application", version.ref = "agp" }
@@ -63,5 +65,6 @@ wearos = [
 ]
 wearos-debug = [
     "androidx-ui-tooling",
+    "androidx-wear-tooling-preview",
     "androidx-ui-test-manifest",
 ]
